+++ strategy-engine/src/main/java/com/tradie/strategy/crypto/dto/CryptoAssetSpec.java
@@ -1,13 +1,15 @@
 package com.tradie.strategy.crypto.dto;
 
+import java.math.BigDecimal;
+
 /**
  * Immutable spec for a crypto asset supported via IBKR Paxos.
  */
 public record CryptoAssetSpec(
         String symbol,
         String name,
-        double minOrderSize,
-        double sizeIncrement,
+        BigDecimal minOrderSize,
+        BigDecimal sizeIncrement,
         int pricePrecision,
         double volatilityMultiplier,
         boolean availableOnIbkr
